model/scripts/requirements.txt
tensorflow~=2.14.0
opencv-python~=4.8.1.78
numpy~=1.26.0
tqdm~=4.66.1
matplotlib~=3.8.0
visualkeras~=0.0.2
pydot~=1.4.2
graphviz~=0.20.1
Pillow~=9.5.0
scikit-image~=0.21.0
keras~=2.14.0
